Moved the DslBindingCustomizer class to the core package in engine
