+++ server-core/engine/src/main/java/io/infectnet/server/engine/content/configuration/DslModule.java
@@ -1,6 +1,6 @@
 package io.infectnet.server.engine.content.configuration;
 
-import io.infectnet.server.engine.content.dsl.DslBindingCustomizer;
+import io.infectnet.server.engine.core.dsl.DslBindingCustomizer;
 import io.infectnet.server.engine.content.dsl.GatherBlock;
 import io.infectnet.server.engine.content.dsl.PlayerStorageDslCustomizer;
 import io.infectnet.server.engine.content.dsl.SelectFilterActionBlock;
