+++ server-core/engine/src/main/java/io/infectnet/server/engine/core/script/selector/SelectorFactory.java
@@ -1,6 +1,6 @@
 package io.infectnet.server.engine.core.script.selector;
 
-import io.infectnet.server.engine.content.dsl.DslBindingCustomizer;
+import io.infectnet.server.engine.core.dsl.DslBindingCustomizer;
 import io.infectnet.server.engine.core.player.Player;
 import io.infectnet.server.engine.core.script.execution.BindingContext;
